pyproject.toml
[tool.poetry]
name = "notes-app"
version = "0.0.0"
description = "notes cli"
authors = ["Lucas Wiman <lucas.wiman@gmail.com>"]
license = "MPL2"

[tool.poetry.dependencies]
python = "^3.10"
pydantic = "^1.9.0"
typer = "^0.4.1"
colorama = "^0.4.4"
PyYAML = "^6.0"
"ruamel.yaml" = "^0.17.21"

[tool.poetry.dev-dependencies]
jupyter = "^1.0.0"
python-dateutil = "^2.8.2"
hypothesis = "^6.43.3"
pytest = "^7.1.1"
parsimonious = "^0.9.0"
pandas = "^1.4.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
